requirements.txt
flask==2.1.2
flask-cors==3.0.10
werkzeug==2.1.2
google-cloud-firestore==2.1.0
